Separated dialogs to its own class
